Struct ogg::CurrentPageValues

struct CurrentPageValues {
    first_page: bool,
    sequence_num: u32,
    segment_cnt: u8,
    cur_pg_lacing: [u8; 255],
    cur_pg_data: Vec<Rc<[u8]>>,
    pck_this_overflow_idx: Option<usize>,
    pck_last_overflow_idx: Option<usize>,
}

Fields
first_page: bool

true if this page is the first one in the logical bitstream

sequence_num: u32

Page counter of the current page Increased for every page

segment_cnt: u8

Points to the first unwritten position in cur_pg_lacing.

cur_pg_lacing: [u8; 255] cur_pg_data: Vec<Rc<[u8]>> pck_this_overflow_idx: Option<usize>

Some(offs), if the last packet couldn't make it fully into this page, and has to be continued in the next page.

offs should point to the first idx in cur_pg_data[last] that should NOT be written in this page anymore.

None if all packets can be written nicely.

pck_last_overflow_idx: Option<usize>

Some(offs), if the first packet couldn't make it fully into the last page, and has to be continued in this page.

offs should point to the first idx in cur_pg_data[0] that hasn't been written.

None if all packets can be written nicely.