Struct ogg::PacketReader

pub struct PacketReader<'a, T: Read + Seek + 'a> {
    rdr: &'a mut T,
    page_infos: HashMap<u32, PageInfo>,
    stream_with_stuff: Option<u32>,
    has_seeked: bool,
}

Reader for packets from an Ogg stream.

This reads codec packets belonging to several different logical streams from one physical Ogg container stream.

If the async feature is activated, and you pass as internal reader a valid implementation of the AdvanceAndSeekBack trait, like the BufReader wrapper, the PacketReader will support async operation, meaning that its internal state doesn't get corrupted if from multiple consecutive reads which it performs, some fail with e.g. the WouldBlock error kind.

Fields
rdr: &'a mut T page_infos: HashMap<u32, PageInfo>

Contains info about all logical streams that

stream_with_stuff: Option<u32>

Contains the stream_serial of the stream that contains some unprocessed packet data. There is always <= 1, bc if there is one, no new pages will be read, so there is no chance for a second to be added None if there is no such stream and one has to read a new page.

has_seeked: bool
Methods

impl<'a, T: Read + Seek + 'a> PacketReader<'a, T>

fn new(rdr: &mut T) -> PacketReader<T>

Constructs a new PacketReader with a given Read.

fn read_packet(self: &mut PacketReader<'a, T>) -> Result<Packet, OggReadError>

Reads a packet, and returns it on success.

fn read_until_pg_header(self: &mut PacketReader<'a, T>)
-> Result<[u8; 27], OggReadError>

Reads until the new page header, and then returns the page header array.

If no new page header is immediately found, it performs a "recapture", meaning it searches for the capture pattern, and if it finds it, it reads the complete first 27 bytes of the header, and returns them.

fn read_ogg_page(self: &mut PacketReader<'a, T>) -> Result<(), OggReadError>

Reads a new Ogg page.

This method reads a new Ogg page.

To support seeking this does not assume that the capture pattern is at the current reader position. Instead it searches until it finds the capture pattern.

fn maybe_advance(&mut self)

fn seek_back(&mut self, len: usize) -> Result<()>

fn seek_bytes(&mut self, pos: SeekFrom) -> Result<u64, Error>

Seeks the underlying reader

Seeks the reader that this PacketReader bases on by the specified number of bytes. All new pages will be read from the new position.

This also flushes all the unread packets in the queue.