Struct ogg::PageInfo

struct PageInfo {
    starts_with_continued: bool,
    first_page: bool,
    last_page: bool,
    last_absgp: u64,
    sequence_num: u32,
    packet_positions: Vec<(u16, u16)>,
    ends_with_continued: bool,
    packet_idx: u8,
    page_body: Vec<u8>,
    last_overlap_pck: Vec<Vec<u8>>,
}

Internal helper struct for PacketReader state

Fields
starts_with_continued: bool

true is the first packet is continued from the page before and false if its a "fresh" one

first_page: bool

true if this page is the first one in the logical bitstream

last_page: bool

true if this page is the last one in the logical bitstream

last_absgp: u64

Absolute granule position of last read page. The codec defines further meaning.

sequence_num: u32

Page counter of last page read

packet_positions: Vec<(u16, u16)>

Packet information: index is number of packet, tuple is (offset, length) of packet if ends_with_continued is true, the last element will contain information about the continued packet

ends_with_continued: bool

true if the last packet is continued in subsequent page(s) false if the last packet has a segment of length < 255 inside this page

packet_idx: u8

The index of the first "unread" packet

page_body: Vec<u8>

Contains the package data

last_overlap_pck: Vec<Vec<u8>>

If there is a residue from previous pages in terms of a package spanning multiple pages, this field contains it. Having this Vec> and not Vec ensures to give us O(n) complexity, not O(n2) for n as number of pages that the packet is contained in.

Methods

impl PageInfo

fn is_first_pck_in_pg(self: &PageInfo) -> bool

Returns true if the first "unread" packet is the first one in the page, false otherwise.

fn is_last_pck_in_pg(self: &PageInfo) -> bool

Returns true if the first "unread" packet is the last one in the page, false otherwise. If the first "unread" packet isn't completed in this page (spans page borders), this returns false.